test: move two_connected_nodes to tests::common and use it in more places
